Organize batch processors into package
