+++ src/batch_processing/batch_processor.py
@@ -3,7 +3,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional
 
 import openai_service
-from base_batch_processor import BaseBatchProcessor
+from batch_processing.base_batch_processor import BaseBatchProcessor
 from document_model import H3Pydantic, PanelPydantic
 from logging_config import get_logger
 from markdown_document import MarkdownDocument
