+++ src/batch_processing/enhanced_batch_processor.py
@@ -3,7 +3,7 @@ from concurrent.futures import ThreadPoolExecutor, as_completed
 from pathlib import Path
 from typing import List, Union
 
-from base_batch_processor import BaseBatchProcessor
+from batch_processing.base_batch_processor import BaseBatchProcessor
 from document_model import H3Pydantic, PanelPydantic
 from logging_config import get_logger
 from markdown_document import MarkdownDocument
